Add assertions
This is to test the correctness of the assumptions made in the program.

Add assertion to method mark, unmark, delete, find to verify if the index is not empty and beyond the bound of user tasklist.
Add assertion method to addTodo, addEvent, and addDeadline to verify that the task description is not empty, and if applicable the time is indicated.

Assertion will allow us to know there is an error in which the program deals with the user input. Using assertion is also a good coding practice.