Delete dead code in class Duke
As GUI is implemented, codes for text-based UI and I/O redirection.

Remove main and run method.
Remove an old constructor for Duke.

The main and run method is no longer needed as class Launch is used to run the program.
As old constructor for Duke, which was used by the main method, is no longer in use.
Delete dead code is a good coding practice.